feat(hooks): add useAppParams and useAppSearchParams hooks
The useAppParams hook is added to the hooks directory. It imports the AppParams type from the '@/types' module and the useParams hook from 'react-router-dom'. This hook is used to retrieve the parameters from the URL.

The useAppSearchParams hook is also added to the hooks directory. It imports the setLimit, setSearch, useAppDispatch, and useAppSelector functions from the '@/redux' module, as well as the useEffect and useSearchParams hooks from 'react' and 'react-router-dom' respectively. This hook is used to manage the search and limit parameters in the application state based on the URL search parameters. It listens for changes in the search and limit values and dispatches actions to update the state accordingly. The updated search and limit values are returned from the hook.